Move theme.
