--- packages/core-react/src/index.ts
@@ -1,3 +1,0 @@
-import createDirectionHelpers from './createDirectionHelpers';
-
-export { createDirectionHelpers };
