+++ packages/core-react/src/index.tsx
@@ -1,0 +1,10 @@
+/**
+ * @copyright   2020, Miles Johnson
+ * @license     https://opensource.org/licenses/MIT
+ */
+
+import createDirectionHelpers from './createDirectionHelpers';
+import createThemeHelpers from './createThemeHelpers';
+import createHOC from './createHOC';
+
+export { createDirectionHelpers, createThemeHelpers, createHOC };
